% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/battingseason.R
\name{battingseason}
\alias{battingseason}
\title{Offensive performance by player season}
\usage{
battingseason(playerID = "parrage01", yearID = "2014")
}
\arguments{
\item{playerID}{- Unique Lahman reference ID}

\item{yearID}{- Season}
}
\value{
Dataframe of batting statistics of \code{playerID} during
\code{yearID}
}
\description{
This function returns a MLB player's offensive performance
statistics from the Lahman baseball database
}

