% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lahman.R
\name{lahman}
\alias{lahman}
\title{Lahman sqlite database.}
\usage{
lahman()
}
\description{
This database is bundled with the package, and contains all data frames
in the lahman2015.sqlite package.
}
\examples{
db <- lahman()
}

