% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/lahman_search.R
\name{lahman_search}
\alias{lahman_search}
\title{Lahman SQLite database search}
\usage{
lahman_search(playerID = "parrage01", nameFirst = "Gerardo",
  nameLast = "Parra", nameGiven = "Enrique", retroID = "parrg001",
  bbrefID = "parrage01")
}
\arguments{
\item{playerID}{- Unique Lahman reference ID}

\item{nameFirst}{- Player First Name}

\item{nameLast}{- Player Last Name}

\item{nameGiven}{- Player Given Name}

\item{retroID}{- Unique Retrosheet reference ID}

\item{bbrefID}{- Unique Baseball-Reference reference ID}
}
\value{
Dataframe of career batting statistics for \code{playerID}
}
\description{
This function returns a queries from the Master table from
the Lahman baseball database to assist in further queries.
}

