% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sabr.R
\docType{package}
\name{sabr}
\alias{sabr}
\alias{sabr-package}
\title{sabr}
\description{
Query and calculate MLB offensive performance statistics
}
\details{
Query offensive performance statistics from the 2015
Lahman MLB SQLite Database
}

