% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaxtonRawls.R
\name{SaxtonRawls}
\alias{SaxtonRawls}
\title{Full Saxton & Rawls Soil Hydraulic Property Calculator}
\usage{
SaxtonRawls(pSand = 50, pClay = 25, pOM = 4)
}
\arguments{
\item{pSand}{Sand content (\%)}

\item{pClay}{Clay content (\%)}

\item{pOM}{Organic matter content (\%)}
}
\value{
Named list of BD, LL15, DUL, SAT, SWCON, and KSAT
}
\description{
Estimates BD, LL15, DUL, SAT, SWCON, and KSAT from sand, clay, and organic matter.
}
