% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaxtonRawls.R
\name{SaxtonRawlsBD}
\alias{SaxtonRawlsBD}
\title{Estimate Bulk Density (BD) using Saxton & Rawls (2006)}
\usage{
SaxtonRawlsBD(pSand = 50, pClay = 25, pOM = 4)
}
\arguments{
\item{pSand}{Sand content (\%)}

\item{pClay}{Clay content (\%)}

\item{pOM}{Organic matter content (\%)}
}
\value{
Bulk density (g/cm³)
}
\description{
Estimate Bulk Density (BD) using Saxton & Rawls (2006)
}
