% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaxtonRawls.R
\name{SaxtonRawlsDUL}
\alias{SaxtonRawlsDUL}
\title{Estimate DUL using Saxton & Rawls (2006)}
\usage{
SaxtonRawlsDUL(pSand = 50, pClay = 25, pOM = 4)
}
\arguments{
\item{pSand}{Sand content (\%)}

\item{pClay}{Clay content (\%)}

\item{pOM}{Organic matter content (\%)}
}
\value{
Drained upper limit (DUL) in \%
}
\description{
Estimate DUL using Saxton & Rawls (2006)
}
