% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaxtonRawls.R
\name{SaxtonRawlsKSAT}
\alias{SaxtonRawlsKSAT}
\title{Estimate Saturated Hydraulic Conductivity (KSAT) using Saxton & Rawls (2006)}
\usage{
SaxtonRawlsKSAT(pSand = 50, pClay = 25, pOM = 4)
}
\arguments{
\item{pSand}{Sand content (\%)}

\item{pClay}{Clay content (\%)}

\item{pOM}{Organic matter content (\%)}
}
\value{
Saturated hydraulic conductivity (mm/day)
}
\description{
Estimate Saturated Hydraulic Conductivity (KSAT) using Saxton & Rawls (2006)
}
