% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaxtonRawls.R
\name{SaxtonRawlsLL15}
\alias{SaxtonRawlsLL15}
\title{Estimate LL15 using Saxton & Rawls (2006)}
\usage{
SaxtonRawlsLL15(pSand = 50, pClay = 25, pOM = 4)
}
\arguments{
\item{pSand}{Sand content (\%)}

\item{pClay}{Clay content (\%)}

\item{pOM}{Organic matter content (\%)}
}
\value{
Lower limit (LL15) in \%
}
\description{
Estimate LL15 using Saxton & Rawls (2006)
}
