% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaxtonRawls.R
\name{SaxtonRawlsSAT}
\alias{SaxtonRawlsSAT}
\title{Estimate SAT using Saxton & Rawls (2006)}
\usage{
SaxtonRawlsSAT(pSand = 50, pClay = 25, pOM = 4)
}
\arguments{
\item{pSand}{Sand content (\%)}

\item{pClay}{Clay content (\%)}

\item{pOM}{Organic matter content (\%)}
}
\value{
Saturation (SAT) in \%
}
\description{
Estimate SAT using Saxton & Rawls (2006)
}
