% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaxtonRawls.R
\name{SaxtonRawlsSWCON}
\alias{SaxtonRawlsSWCON}
\title{Estimate SWCON using Saxton & Rawls (2006)}
\usage{
SaxtonRawlsSWCON(pSand = 50, pClay = 25, pOM = 4)
}
\arguments{
\item{pSand}{Sand content (\%)}

\item{pClay}{Clay content (\%)}

\item{pOM}{Organic matter content (\%)}
}
\value{
SWCON (drainage coefficient) in \%
}
\description{
Estimate SWCON using Saxton & Rawls (2006)
}
