% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/woba_guts.R
\docType{data}
\name{woba_guts}
\alias{woba_guts}
\title{wOBA guts table}
\format{An object of class \code{"data.frame"}.}
\source{
\href{http://www.fangraphs.com/guts.aspx?type=cn}{Fangraphs.com}
}
\usage{
woba_guts
}
\description{
This dataframe contains the yearly wOBA guts weights
from Fangraphs (http://www.fangraphs.com/guts.aspx?type=cn)
}
\references{
Andrew Dolphin, Mitchel Lichtman, and Tom Tango (2007)
The Book: Playing the Percentages in Baseball
}
\keyword{datasets}

