% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/woba_season.R
\name{woba_season}
\alias{woba_season}
\title{wOBA performance by player season}
\usage{
woba_season(playerID = "parrage01", yearID = "2014")
}
\arguments{
\item{playerID}{- Unique Lahman reference ID}

\item{yearID}{- Season}
}
\value{
Print wOBA of \code{playerID} during \code{yearID}
}
\description{
This function returns a MLB player's weight On-Base Average
statistic using the Lahman baseball database and Fangraphs weights
}

