% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wRAA_season.R
\name{wRAA_season}
\alias{wRAA_season}
\title{wRAA performance by player season}
\usage{
wRAA_season(playerID = NULL, yearID = NULL, ...)
}
\arguments{
\item{playerID}{- Unique Lahman reference ID}

\item{yearID}{- Season}
}
\value{
Print wRAA of \code{playerID} during \code{yearID}
}
\description{
This function returns a MLB player's weighted Runs Above Average
statistic using the Lahman baseball database statistics/park factors
and Tangotiger's formula for calculating wOBA weights/scales
}
\examples{
wRAA_season("braunry02","2015")
}
\references{
Andrew Dolphin, Mitchel Lichtman, and Tom Tango (2007)
            The Book: Playing the Percentages in Baseball
}

