% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wRC_season.R
\name{wRC_season}
\alias{wRC_season}
\title{wRC performance by player season}
\usage{
wRC_season(playerID = NULL, yearID = NULL, ...)
}
\arguments{
\item{playerID}{- Unique Lahman reference ID}

\item{yearID}{- Season}
}
\value{
Print wRC of \code{playerID} during \code{yearID}
}
\description{
This function returns a MLB player's weighted Runs Created 
statistic using the Lahman baseball database statistics/park factors
and Tangotiger's formula for calculating wOBA weights/scales
}
\examples{
wRC_season("molitpa01","1991")
}
\references{
Andrew Dolphin, Mitchel Lichtman, and Tom Tango (2007)
            The Book: Playing the Percentages in Baseball
}

