% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wRCp_season.R
\name{wRCp_season}
\alias{wRCp_season}
\title{wRCp performance by player season}
\usage{
wRCp_season(playerID = NULL, yearID = NULL, ...)
}
\arguments{
\item{playerID}{- Unique Lahman reference ID}

\item{yearID}{- Season}
}
\value{
Print wRC+ of \code{playerID} during \code{yearID}
}
\description{
This function returns a MLB player's weighted Runs Created Plus
statistic using the Lahman baseball database statistics/park factors
and Tangotiger's formula for calculating wOBA weights/scales
}
\references{
Andrew Dolphin, Mitchel Lichtman, and Tom Tango (2007)
            The Book: Playing the Percentages in Baseball
}

