% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingcareer.R
\name{battingcareer}
\alias{battingcareer}
\title{Offensive seasonal performance by player career}
\usage{
battingcareer(playerID = NULL)
}
\arguments{
\item{playerID}{- Unique Lahman reference ID}
}
\value{
Dataframe of career batting statistics for \code{playerID}
}
\description{
This function returns a MLB player's career performance
statistics from the Lahman baseball database
}

