% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battingseason.R
\name{battingseason}
\alias{battingseason}
\title{Offensive performance by player season}
\usage{
battingseason(playerID = NULL, yearID = NULL)
}
\arguments{
\item{playerID}{- Unique Lahman reference ID}

\item{yearID}{- Season}
}
\value{
Dataframe of batting statistics of \code{playerID} during
\code{yearID}
}
\description{
This function returns a MLB player's offensive performance
statistics from the Lahman baseball database
}
\examples{
battingseason("parrage01","2014")
}

