% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lahman_search.R
\name{lahman_search}
\alias{lahman_search}
\title{Lahman SQLite database search}
\usage{
lahman_search(playerID = NULL, nameFirst = NULL, nameLast = NULL,
  nameGiven = NULL, retroID = NULL, bbrefID = NULL)
}
\arguments{
\item{playerID}{- Unique Lahman reference ID}

\item{nameFirst}{- Player First Name}

\item{nameLast}{- Player Last Name}

\item{nameGiven}{- Player Given Name}

\item{retroID}{- Unique Retrosheet reference ID}

\item{bbrefID}{- Unique Baseball-Reference reference ID}
}
\value{
Dataframe of career batting statistics for \code{playerID}
}
\description{
This function returns a queries from the Master table from
the 2015 Lahman baseball database to assist in further queries.
}
\examples{
lahman_search(nameFirst = "Robin")
lahman_search(nameLast = "Fielder")
lahman_search(nameGiven = "Russell Oles")
lahman_search(retroID = "molip001")
lahman_search(bbrefID = "yostne01")
}

