% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/sabr.R
\docType{package}
\name{sabr}
\alias{sabr}
\alias{sabr-package}
\title{sabr}
\description{
Query and calculate MLB offensive performance statistics
}
\details{
Query offensive performance statistics from the
Lahman MLB SQL Database and calculate wOBA using the wOBA
constants (http://www.fangraphs.com/guts.aspx?type=cn)
}

