% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/woba_guts.R
\name{woba_guts}
\alias{woba_guts}
\title{Retrieves wOBA guts table}
\usage{
woba_guts()
}
\value{
Dataframe of woba weights from Fangraphs
(http://www.fangraphs.com/guts.aspx?type=cn)
}
\description{
This function returns the yearly wOBA guts weights
from Fangraphs (http://www.fangraphs.com/guts.aspx?type=cn)
}

