% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woba_season.R
\name{woba_season}
\alias{woba_season}
\title{wOBA performance by player season}
\usage{
woba_season(playerID = NULL, yearID = NULL, ...)
}
\arguments{
\item{playerID}{- Unique Lahman reference ID}

\item{yearID}{- Season}
}
\value{
Print wOBA of \code{playerID} during \code{yearID}
}
\description{
This function returns a MLB player's weight On-Base Average
statistic using the Lahman baseball database and Fangraphs weights
}
\references{
Andrew Dolphin, Mitchel Lichtman, and Tom Tango (2007)
            The Book: Playing the Percentages in Baseball
}

